requirements.txt
streamlit==1.41.1
pandas==2.2.2
numpy==1.26.4
plotly==5.19.0
supabase==2.3.4
gotrue==2.8.1
httpx==0.25.2
xlsxwriter==3.2.0
streamlit-javascript==0.1.5
python-dotenv==1.0.1
argon2-cffi==23.1.0
pytz==2024.1pytest==9.0.2
pytest-cov==7.0.0
pydantic==2.12.0